Reorganize groups code
