+++ apps/prairielearn/src/components/GroupWorkInfoContainer.tsx
@@ -1,7 +1,7 @@
 import { html } from '@prairielearn/html';
 
 import type { GroupConfig } from '../lib/db-types.js';
-import type { GroupInfo } from '../lib/groups.js';
+import type { GroupInfo } from '../lib/groups.shared.js';
 import { idsEqual } from '../lib/id.js';
 
 import { Modal } from './Modal.js';
