+++ apps/prairielearn/src/components/QuestionContainer.tsx
@@ -18,7 +18,8 @@ import type {
   User,
   Variant,
 } from '../lib/db-types.js';
-import { type GroupInfo, getRoleNamesForUser } from '../lib/groups.js';
+import { getRoleNamesForUser } from '../lib/groups.js';
+import type { GroupInfo } from '../lib/groups.shared.js';
 import { idsEqual } from '../lib/id.js';
 import type { IssueRenderData } from '../lib/question-render.types.js';
 import type { UntypedResLocals } from '../lib/res-locals.types.js';
