+++ apps/prairielearn/src/pages/studentAssessment/studentAssessment.html.ts
@@ -4,7 +4,7 @@ import { html, unsafeHtml } from '@prairielearn/html';
 import { GroupWorkInfoContainer } from '../../components/GroupWorkInfoContainer.js';
 import { PageLayout } from '../../components/PageLayout.js';
 import { type Assessment, type GroupConfig, type User } from '../../lib/db-types.js';
-import { type GroupInfo } from '../../lib/groups.js';
+import type { GroupInfo } from '../../lib/groups.shared.js';
 import type { ResLocalsForPage } from '../../lib/res-locals.js';
 
 export function StudentAssessment({
